調整AuthorServiceTest、新增AuthorControllerTest 調整所有的控制器的api路徑以便更好的配合RESTFulFul

+++ src/main/java/com/oovetest/webDemo/author/service/AuthorService.java
@@ -112,6 +112,9 @@ public class AuthorService {
     }
 
     public void deleteAuthorById(@NonNull Long authorId) {
+        if (!authorRepository.existsById(authorId)) {
+            throw new NotFoundException("查無此作者ID");
+        }
         authorRepository.deleteById(authorId);
     }
 }
